[feat/#175] :recycle: Refactor: RoomUsersStatus를 domain 디렉토리로 분리
