+++ chatty-be/src/main/java/com/chatty/chatty/quizroom/service/GameService.java
@@ -1,6 +1,6 @@
 package com.chatty.chatty.quizroom.service;
 
-import com.chatty.chatty.quizroom.controller.dto.RoomUsersStatus;
+import com.chatty.chatty.quizroom.domain.RoomUsersStatus;
 import com.chatty.chatty.quizroom.repository.RoomUsersStatusRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -10,7 +10,7 @@ import org.springframework.stereotype.Service;
 public class GameService {
 
     private final RoomUsersStatusRepository roomUsersStatusRepository;
-    
+
     public RoomUsersStatus joinRoom(Long roomId, Long userId) {
         return roomUsersStatusRepository.addUserToRoom(roomId, userId);
     }
