+++ chatty-be/src/main/java/com/chatty/chatty/quizroom/service/QuizRoomService.java
@@ -10,8 +10,8 @@ import com.chatty.chatty.quizroom.controller.dto.MakeRoomRequest;
 import com.chatty.chatty.quizroom.controller.dto.MakeRoomResponse;
 import com.chatty.chatty.quizroom.controller.dto.RoomDetailResponse;
 import com.chatty.chatty.quizroom.controller.dto.RoomQuizResponse;
-import com.chatty.chatty.quizroom.entity.QuizRoom;
-import com.chatty.chatty.quizroom.entity.Status;
+import com.chatty.chatty.quizroom.domain.entity.QuizRoom;
+import com.chatty.chatty.quizroom.domain.entity.Status;
 import com.chatty.chatty.quizroom.exception.QuizRoomException;
 import com.chatty.chatty.quizroom.repository.QuizRoomRepository;
 import com.chatty.chatty.quizroom.repository.RoomUsersStatusRepository;
